wip: abstract prepareColorGroups function for re-use
