+++ packages/theme-wizard-app/src/components/wizard-color-system-preview/index.ts
@@ -5,14 +5,11 @@ import '../wizard-table-scroller';
 
 import { arrayFromCommaList } from '@nl-design-system-community/clippy-components/lib/converters';
 import { safeCustomElement } from '@nl-design-system-community/clippy-components/lib/decorators';
-import { stringifyColor } from '@nl-design-system-community/design-tokens-schema';
-import { type ColorToken as ColorTokenType } from '@nl-design-system-community/design-tokens-schema';
 import { LitElement, html } from 'lit';
 import { property, state } from 'lit/decorators.js';
 import type Theme from '../../lib/Theme';
 import { themeContext } from '../../contexts/theme';
-import { resolveColorValue } from '../wizard-colorscale-input';
-import { countUsagePerToken } from '../wizard-style-guide/utils';
+import { countUsagePerToken, prepareColorGroups } from '../wizard-style-guide/utils';
 
 const tag = 'wizard-color-system-preview';
 
@@ -33,26 +30,6 @@ export class WizardColorSystemPreview extends LitElement {
   @property({ converter: arrayFromCommaList })
   groups: string[] = [];
 
-  #prepareColorGroups(colors: Record<string, unknown>, tokenUsage: Map<string, string[]>): ColorGroup[] {
-    return Object.entries(colors)
-      .filter(([key]) => !key.includes('inverse') && !key.includes('transparent'))
-      .filter(([, value]) => typeof value === 'object' && value !== null)
-      .map(([key, value]) => {
-        const colorEntries = Object.entries(value as Record<string, unknown>)
-          .filter(([, token]) => typeof token === 'object' && token !== null && '$value' in token)
-          .map(([colorKey, token]) => {
-            const color = resolveColorValue(token as ColorTokenType);
-            const displayValue = color ? stringifyColor(color) : '#000';
-            const tokenId = `basis.color.${key}.${colorKey}`;
-            const usage = tokenUsage.get(tokenId) || [];
-            const usageCount = usage.length;
-            return { colorKey, displayValue, tokenId, usage, usageCount };
-          })
-          .filter(({ displayValue }) => displayValue !== null);
-        return { colorEntries, key };
-      });
-  }
-
   override connectedCallback() {
     super.connectedCallback();
 
@@ -60,7 +37,7 @@ export class WizardColorSystemPreview extends LitElement {
     const colors = basis['color'] as Record<string, unknown>;
     const tokenUsage = countUsagePerToken(this.theme.tokens);
 
-    const colorTokenGroups = this.#prepareColorGroups(colors, tokenUsage);
+    const colorTokenGroups = prepareColorGroups(colors, tokenUsage);
     this.#visibleTokenGroups = this.#filterColorGroups(this.groups, colorTokenGroups);
   }
